move components into `lib`
