+++ packages/skeleton-react/src/lib/components/Accordion/schema.json
