+++ packages/skeleton-react/src/lib/components/Accordion/types.ts
