+++ packages/skeleton-react/src/lib/components/AppBar/schema.json
