+++ packages/skeleton-react/src/lib/components/AppBar/types.ts
