+++ packages/skeleton-react/src/lib/components/Avatar/schema.json
