+++ packages/skeleton-react/src/lib/components/Avatar/types.ts
