+++ packages/skeleton-react/src/lib/components/Progress/schema.json
