+++ packages/skeleton-react/src/lib/components/Progress/types.ts
