+++ packages/skeleton-react/src/lib/components/Tabs/schema.json
