+++ packages/skeleton-react/src/lib/index.ts
@@ -1,16 +1,16 @@
-export { default as TestReact } from "../components/Test/Test.jsx";
+export { default as TestReact } from "./components/Test/Test.js";
 
 // Components
-export { Accordion } from "../components/Accordion/Accordion.jsx";
-export { Avatar } from "../components/Avatar/Avatar.jsx";
-export { AppBar, ToolBar } from "../components/AppBar/AppBar.jsx";
-export { Progress } from "../components/Progress/Progress.jsx";
-export { Tabs } from "../components/Tabs/Tabs.jsx";
+export { Accordion } from "./components/Accordion/Accordion.js";
+export { Avatar } from "./components/Avatar/Avatar.js";
+export { AppBar, ToolBar } from "./components/AppBar/AppBar.js";
+export { Progress } from "./components/Progress/Progress.js";
+export { Tabs } from "./components/Tabs/Tabs.js";
 
 // Typescript Schema
 // REMINDER: append `Schema` as the suffix
-export { default as AccordionSchema } from "../components/Accordion/schema.json";
-export { default as AvatarSchema } from "../components/Avatar/schema.json";
-export { default as AppBarSchema } from "../components/AppBar/schema.json";
-export { default as ProgressSchema } from "../components/Progress/schema.json";
-export { default as TabsSchema } from "../components/Tabs/schema.json";
+export { default as AccordionSchema } from "./components/Accordion/schema.json";
+export { default as AvatarSchema } from "./components/Avatar/schema.json";
+export { default as AppBarSchema } from "./components/AppBar/schema.json";
+export { default as ProgressSchema } from "./components/Progress/schema.json";
+export { default as TabsSchema } from "./components/Tabs/schema.json";
