+++ packages/skeleton-react/src/routes/components/test.tsx
@@ -1,5 +1,5 @@
-import Test from "../../components/Test/Test";
+import Test from "../../lib/components/Test/Test.js";
 
 export function Component() {
-  return <Test />;
+	return <Test />;
 }
